Collapse <patchwork-view> legacy mode back onto a single DOM node
The 1.0 reshape of <patchwork-view> into a wrapper that delegated to an
inner <patchwork-view-legacy> child broke event-driven coordination
between bootloaders that dispatchEvent on the wrapper and tools that
addEventListener on their own mount host (the child): events dispatched
on the wrapper never bubbled _down_ to the child, and listeners on the
child could stopPropagation before ancestors saw the event.

Host the doc-url / tool-id render lifecycle directly on <patchwork-view>
via a shared LegacyImpl class, so the dispatch target and the tool's
mount host are the same node. Delete <patchwork-view-legacy>,
registerPatchworkViewLegacyElement, PatchworkViewLegacyElement, and its
JSX intrinsic — they're no longer needed once the split is gone.

Roll @inkandswitch/patchwork-elements back from 1.0.3 to 0.0.8 (the
1.0.x line is being unpublished from npm) and add changesets so the
next release rolls forward to 0.1.0 (elements) and 0.2.4 (bootloader,
doc-only patch).
