+++ core/elements/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@inkandswitch/patchwork-elements",
-  "version": "1.0.3",
+  "version": "0.0.8",
   "author": "Ink & Switch",
   "type": "module",
   "description": "",
