+++ core/elements/src/index.ts
@@ -1,15 +1,13 @@
 export * from "./patchwork-view.js";
-export * from "./patchwork-view-legacy.js";
 export * from "./events.js";
 
 import type {} from "react";
 import type {} from "solid-js";
 
 /**
- * Attributes for the legacy mode of `<patchwork-view>`: the wrapper
- * delegates to an inner `<patchwork-view-legacy>` driven by `doc-url`
- * and `tool-id`. The component-mode attributes are typed as `never` so
- * that mixing the two modes in JSX is a compile-time error.
+ * Attributes for the legacy mode of `<patchwork-view>`, driven by
+ * `doc-url` and `tool-id`. The component-mode attributes are typed as
+ * `never` so that mixing the two modes in JSX is a compile-time error.
  */
 type PatchworkViewLegacyAttrs = {
   "doc-url"?: string;
@@ -19,10 +17,9 @@ type PatchworkViewLegacyAttrs = {
 };
 
 /**
- * Attributes for the component mode of `<patchwork-view>`: the wrapper
- * mounts a registered `patchwork:component` plugin in place. The
- * legacy-mode attributes are typed as `never` for the same reason as
- * above.
+ * Attributes for the component mode of `<patchwork-view>`: mounts a
+ * registered `patchwork:component` plugin in place. The legacy-mode
+ * attributes are typed as `never` for the same reason as above.
  */
 type PatchworkViewComponentAttrs = {
   "doc-url"?: never;
@@ -40,11 +37,6 @@ type PatchworkViewAttrs =
   | PatchworkViewLegacyAttrs
   | PatchworkViewComponentAttrs;
 
-type PatchworkViewLegacyElementAttrs = {
-  "doc-url": string;
-  "tool-id"?: string | null;
-};
-
 declare module "react" {
   export namespace JSX {
     interface IntrinsicElements {
@@ -52,10 +44,6 @@ declare module "react" {
         React.HTMLAttributes<HTMLElement>,
         HTMLElement
       > & { class?: string } & PatchworkViewAttrs;
-      "patchwork-view-legacy": React.DetailedHTMLProps<
-        React.HTMLAttributes<HTMLElement>,
-        HTMLElement
-      > & { class?: string } & PatchworkViewLegacyElementAttrs;
     }
   }
 }
@@ -64,7 +52,6 @@ declare module "solid-js" {
   namespace JSX {
     interface IntrinsicElements {
       "patchwork-view": PatchworkViewAttrs;
-      "patchwork-view-legacy": PatchworkViewLegacyElementAttrs;
     }
   }
 }
@@ -78,10 +65,6 @@ declare global {
         React.HTMLAttributes<HTMLElement>,
         HTMLElement
       > & { class?: string } & PatchworkViewAttrs;
-      "patchwork-view-legacy": React.DetailedHTMLProps<
-        React.HTMLAttributes<HTMLElement>,
-        HTMLElement
-      > & { class?: string } & PatchworkViewLegacyElementAttrs;
     }
   }
 }
